Fix typo in bulalo.html
As pot was already used in step 1 of
ube-jam.html, the ube can already fit and pot is
not in use until 2 steps later. Change kitchenware
to fix consistency

Cutting and chopping actions were called for in
the recipe pages' "How to make it" list but no
specification using what tool. Add knives
and cutting boards